third_party/blink/web_tests/platform/generic/plugins/plugin-scriptable-expected.txt

This tests checks that hasOwnProperty return true on plugin's getters that return undefined

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new ReferenceObject().hasOwnProperty("testGetUndefined") is true
PASS document.getElementById("plugin").hasOwnProperty("testGetUndefined") is true
PASS successfullyParsed is true

TEST COMPLETE
